refactor: 배달부 캐릭터 관련 상수 constants 폴더로 분리
